requirements.txt
numpy
mysql-connector-python
Pillow
opencv-contrib-python
python-dotenv
